+++ src/Resources/CheckoutSessionsResource.php
@@ -10,6 +10,7 @@ use Saloon\Http\BaseResource;
 use Subster\PhpSdk\DataObjects\CheckoutSessionData;
 use Subster\PhpSdk\DataObjects\CheckoutSessionStatusData;
 use Subster\PhpSdk\DataObjects\CreateCheckoutSessionData;
+use Subster\PhpSdk\Requests\CancelCheckoutSessionRequest;
 use Subster\PhpSdk\Requests\CreateCheckoutSessionRequest;
 use Subster\PhpSdk\Requests\GetCheckoutSessionRequest;
 
@@ -34,6 +35,17 @@ class CheckoutSessionsResource extends BaseResource
     {
         return $this->connector->send(
             new GetCheckoutSessionRequest($session)
+        )->dtoOrFail();
+    }
+
+    /**
+     * @throws FatalRequestException
+     * @throws RequestException
+     */
+    public function cancel(string $session): CheckoutSessionStatusData
+    {
+        return $this->connector->send(
+            new CancelCheckoutSessionRequest($session)
         )->dtoOrFail();
     }
 }